% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp2022extra.R
\docType{package}
\name{wpp2022extra-package}
\alias{wpp2022extra-package}
\alias{wpp2022extra}
\title{Additional Datasets from the World Population Prospects 2022}
\description{
Provides additional data to the wpp2022 package from the United Nation's World Population Prospects 2022.
}
